+++ SlowpokeHubs/BodyFacade.cs
@@ -25,7 +25,7 @@ namespace SlowpokeHubs
         public ShapeFacade Shape { get; set; }
         //Points per second
         public int Speed { get; set; }
-        public int LastProcessedCommandId { get; set; }
+        public long LastProcessedCommandId { get; set; }
 
         public static BodyFacade FromBody(Body body)
         {
